requirements.txt
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
prettytable==3.12.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
wcwidth==0.2.13
